LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/events/invalid-uncompiled-raw-handler-compiled-late.window-expected.txt

CONSOLE MESSAGE: SyntaxError: Parser error
CONSOLE MESSAGE: SyntaxError: Parser error

FAIL Invalid uncompiled raw handlers should only be compiled when about to call them assert_array_equals: lengths differ, expected array ["click 1", "error", "click 2"] length 3, got ["click 1", "error", "click 2", "error"] length 4
